feat: add login verification in all pages
